refactor: folder renamed
- updated folder from diagnostic to diagnostics

- updated the cmake to include the files
