+++ toolchain/CMakeLists.txt
@@ -3,6 +3,11 @@
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 # Collect source files for each component
+file(GLOB_RECURSE DIAGNOSTICS_SOURCES
+    "${CMAKE_CURRENT_SOURCE_DIR}/diagnostics/*.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/diagnostics/*.hpp"
+)
+
 file(GLOB_RECURSE SOURCE_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/source/*.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/source/*.hpp"
@@ -30,6 +35,7 @@ file(GLOB_RECURSE CHECKER_SOURCES
 
 # Add them to parent scope
 set(TOOLCHAIN_SOURCES
+    ${DIAGNOSTICS_SOURCES}
     ${SOURCE_SOURCES}
     ${LEX_SOURCES}
     ${AST_SOURCES}
@@ -40,6 +46,7 @@ set(TOOLCHAIN_SOURCES
 
 # Add include directories
 target_include_directories(zivc PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/diagnostics
     ${CMAKE_CURRENT_SOURCE_DIR}/source
     ${CMAKE_CURRENT_SOURCE_DIR}/lex
     ${CMAKE_CURRENT_SOURCE_DIR}/ast
