+++ toolchain/diagnostics/diagnostic_consumer.cpp
