+++ toolchain/diagnostics/diagnostic_consumer.hpp
