+++ toolchain/diagnostics/diagnostic_emitter.hpp
