+++ toolchain/diagnostics/diagnostic_kind.cpp
