+++ toolchain/diagnostics/diagnostic_kind.hpp
