+++ toolchain/diagnostics/diagnostic_kind_registry.def
