+++ zivc/toolchain/commands/lex_command.cpp
@@ -7,7 +7,7 @@
 #include "toolchain/ast/tree.hpp"
 #include "toolchain/parser/parser.hpp"
 #include "toolchain/ast/tree.hpp"
-#include "toolchain/diagnostic/diagnostic_consumer.hpp"
+#include "toolchain/diagnostics/diagnostic_consumer.hpp"
 
 namespace ziv::cli::toolchain {
